feat(kg-bayesian): Task 4 — Laplace approximation, NUTS sampling, and subgraph extraction
Implements inference.py with three public callables:
- extract_inference_subgraph: BFS subgraph around target nodes
- laplace_approximate: MAP via L-BFGS-B + Hessian covariance (LaplaceResult)
- sample_posterior: NUTS via NumPyro MCMC (NUTSResult + BayesianDiagnostics)

Also fixes diagnostics.py: arviz summary r_hat/ess_bulk columns return as
object dtype (string) in this arviz version — cast to float before comparison.
All 6 new tests pass; 15 existing bayesian tests unaffected.

+++ autoreview/knowledge_graph/bayesian/diagnostics.py
@@ -79,9 +79,10 @@ def compute_mcmc_diagnostics(
     idata = az.from_numpyro(mcmc)
 
     # Per-edge summary: R-hat, ESS
+    # arviz may return summary columns as object dtype (string) — cast to float
     summary = az.summary(idata, var_names=["truth"])
-    r_hats = summary["r_hat"].values
-    ess_bulk = summary["ess_bulk"].values
+    r_hats = summary["r_hat"].values.astype(float)
+    ess_bulk = summary["ess_bulk"].values.astype(float)
 
     # Divergences
     n_divergences = 0
